fix: height type was uint32 instead of int64 (dashevo/dapi-grpc#123)

+++ packages/dapi-grpc/clients/platform/v0/objective-c/Platform.pbobjc.h
@@ -78,7 +78,7 @@ typedef GPB_ENUM(ResponseMetadata_FieldNumber) {
 
 @interface ResponseMetadata : GPBMessage
 
-@property(nonatomic, readwrite) uint32_t height;
+@property(nonatomic, readwrite) int64_t height;
 
 @property(nonatomic, readwrite) uint32_t coreChainLockedHeight;
